+++ changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.2.0 (2026, september 9)
+
+### Compatibility with other patchs
+- In version 1.0.1 I made an extra patch to ensure compatibility with "Visual Improvement". I've been testing without the extra patch and the game seems to work fine, so for now I won't be making the secondary patch. To use REharmonized + Visual Improvement, follow these steps:
+  - STEP 1: Apply REharmonized patch first.
+  - STEP 2: Apply "Visual Improvement" (Pemburu Vampir) to the result of STEP 1.
+
+- Applying REharmonized first allows other patches to be applied afterwards, but sometimes it won't work 100%. **I'll keep an eye on community feedback.**
+
+### Music
+- Old enemy: WSPursuer's version, cleaner melody.
+- Boss battle: Good Knight Productions' version, cleaner melody.
+- Prologue 2: some parts of the song sounded a bit saturated. I've made adjustments to the script to manually control volume normalization when needed, and I did so in this case.
+
 ## 1.0.1 (2026, september 8)
 - Second patch REharmonized + Visual Improvement by Pemburu Vampir
   - Apply first Visual Improvement 1.2.7, then apply REharmonized